+++ src/sinusstring.h
@@ -21,7 +21,7 @@ protected:
 	int m_nSinPos;
 	int m_nAmplifier;
 	int m_nY;
-	float m_nX;
+	double m_nX;
 	int m_nStartPos;
 	int	m_nStrLen;
 	char *m_pchString;
